+++ errors.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/Azure/azure-amqp-common-go/v2/rpc"
+	"github.com/Azure/azure-amqp-common-go/v3/rpc"
 )
 
 type (