+++ message.go
@@ -29,10 +29,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Azure/azure-amqp-common-go/v2/uuid"
+	"github.com/Azure/azure-amqp-common-go/v3/uuid"
 	"github.com/devigned/tab"
 	"github.com/mitchellh/mapstructure"
-	"pack.ag/amqp"
+	"github.com/Azure/go-amqp"
 )
 
 type (