+++ sender.go
@@ -24,13 +24,13 @@ package servicebus
 
 import (
 	"context"
-	common "github.com/Azure/azure-amqp-common-go/v2"
+	common "github.com/Azure/azure-amqp-common-go/v3"
 	"sync"
 	"time"
 
-	"github.com/Azure/azure-amqp-common-go/v2/uuid"
+	"github.com/Azure/azure-amqp-common-go/v3/uuid"
 	"github.com/devigned/tab"
-	"pack.ag/amqp"
+	"github.com/Azure/go-amqp"
 )
 
 type (