+++ session.go
@@ -27,9 +27,9 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/Azure/azure-amqp-common-go/v2/uuid"
+	"github.com/Azure/azure-amqp-common-go/v3/uuid"
 	"github.com/devigned/tab"
-	"pack.ag/amqp"
+	"github.com/Azure/go-amqp"
 )
 
 type (